requirements.txt
streamlit==1.34.0
pymupdf==1.23.5
chromadb==0.4.18
cloudinary==1.36.0
pillow==10.0.1
imagehash==4.3.1
tiktoken==0.5.1
python-dotenv==1.0.0
google-generativeai==0.3.1